Add new features and CSS animations to app.vue
The update includes adding a music file to play when an event is triggered. Also introduced, are some new functional elements with respective CSS animations to the 'app.vue' file. Additionally, the configuration file 'tailwind.config.js' is slightly altered. Lastly, an empty line is also added in the 'nuxt.config.ts' file.